requirements.txt
flask
python-dotenv
flask-mysql
sendgrid